+++ src/components/Callout/CalloutMenu.js
@@ -1,25 +1,48 @@
 import React from 'react'
 import MoreIcon from 'react-icons/lib/md/more-vert'
 import Callout from './index'
-import uuid from 'uuid/v4'
 import { plainButtonRule } from '../Button'
 
-const CalloutMenu = ({ menu, styles }) => {
+const hasAncestor = (node, predicate) => {
+  if (predicate(node)) {
+    return true
+  }
+  if (node.parentNode) {
+    return hasAncestor(node.parentNode, predicate)
+  }
+  return false
+}
+
+const CalloutMenu = ({ menu, styles, icon, leftAligned }) => {
   const [showMenu, setMenu] = React.useState(false)
   const toggleRef = React.useRef()
 
+  const handleClick = e => {
+    if (!hasAncestor(e.target, node => node === toggleRef.current)) {
+      setMenu(false)
+    }
+  }
+
+  React.useEffect(() => {
+    if (!showMenu) return
+    window.addEventListener('click', handleClick)
+    return () => {
+      window.removeEventListener('click', handleClick)
+    }
+  }, [showMenu])
+
   return menu ? (
-    <div id={uuid()} ref={toggleRef} {...styles}>
+    <div ref={toggleRef} {...styles}>
       <button
         {...plainButtonRule}
         onClick={() => {
           setMenu(!showMenu)
         }}
       >
-        <MoreIcon width='calc(1em + 7px)' />
+        {icon || <MoreIcon width='calc(1em + 7px)' />}
       </button>
       {showMenu && (
-        <Callout toggleRef={toggleRef} onClose={setMenu(false)}>
+        <Callout onClose={() => setMenu(false)} leftAligned={leftAligned}>
           {menu}
         </Callout>
       )}
