+++ src/index.js
@@ -221,9 +221,7 @@ ReactDOM.render(
               path: '/callout',
               title: 'Callout',
               imports: {
-                Callout: require('./components/Callout'),
-                MoreIcon: require('react-icons/lib/md/more-vert'),
-                plainButtonRule: require('./components/Button').plainButtonRule
+                CalloutMenu: require('./components/Callout/CalloutMenu')
               },
               src: require('./components/Callout/docs.md')
             }
